Squashed commit of the following:
commit 42feb619d77b46b91f6d5868bccb4208598f4830
Author: Melina Morales <melina.morales@mi.unc.edu.ar>
Date:   Fri Oct 3 16:16:41 2025 -0300

    feature: healing potion

+++ Assets/Scripts/Enemy.cs
@@ -13,7 +13,7 @@ public abstract class Enemy : HealthSystem
     public float distanceToStop = 5f;
 
     [Header("Attack")]
-    [SerializeField] protected float _damage;
+    [SerializeField] protected int _damage;
     [SerializeField] protected float _nextAttackTime;
     [SerializeField] protected float _attackCooldown;
